+++ src/thermo/GibbsExcessVPSSTP.cpp
@@ -257,7 +257,7 @@ void GibbsExcessVPSSTP::getActivityCoefficients(doublereal* const ac) const
     getLnActivityCoefficients(ac);
 
     // Protect against roundoff when taking exponentials
-    for (int k = 0; k < m_kk; k++) {
+    for (size_t k = 0; k < m_kk; k++) {
         if (ac[k] > 700.) {
             ac[k] = exp(700.0);
         } else if (ac[k] < -700.) {
